+++ src/node/start.ts
@@ -1,12 +1,17 @@
-import { createServer as createViteServer } from 'vite';
-import { resolveConfig } from './config.js';
-import { UserConfig } from './types.js';
-import { docit } from './plugins/index.js';
+import { createServer as createViteServer } from "vite";
+import { resolveConfig } from "./config.js";
+import { UserConfig } from "./types.js";
+import { docit } from "./plugins/index.js";
 
 export const start = async (userConfig: UserConfig) => {
-  const config = await resolveConfig(userConfig, 'start');
+  const config = await resolveConfig(userConfig, "start");
   return createViteServer({
-    base: '/',
+    base: "/",
     plugins: [await docit(config)],
-  });
+  })
+    .then((s) => s.listen())
+    .then((s) => {
+      s.printUrls();
+      return s;
+    });
 };