+++ CHANGELOG.md
@@ -12,6 +12,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## [0.6.6] - Unreleased
+
+### Fix
+- fix `types` is missing in `package.json`
+- fix react-frame-component does not provide an export named `default`
+
+### Added
+- add `exports` in `package.json`
+
+### Changed
+- `start` api will start server
+
 ## [0.6.5] - 2022-03-30
 
 ### Added